Correct errors in the flow algorithm

+++ source/shift_x.hpp
@@ -10,7 +10,7 @@ namespace tarsier {
     template <typename Event, typename HandleEvent>
     class shift_x {
         public:
-        shift_x(uint16_t width, uint16_t shift, HandleEvent handle_event) :
+        shift_x(uint16_t width, int32_t shift, HandleEvent handle_event) :
             _width(width),
             _shift(shift),
             _handle_event(std::forward<HandleEvent>(handle_event)) {}
@@ -31,13 +31,13 @@ namespace tarsier {
 
         protected:
         const uint16_t _width;
-        const uint16_t _shift;
+        const int32_t _shift;
         HandleEvent _handle_event;
     };
 
     /// make_shift_x creates a shift_x from a functor.
     template <typename Event, typename HandleEvent>
-    shift_x<Event, HandleEvent> make_shift_x(uint16_t width, uint16_t shift, HandleEvent handle_event) {
+    shift_x<Event, HandleEvent> make_shift_x(uint16_t width, int32_t shift, HandleEvent handle_event) {
         return shift_x<Event, HandleEvent>(width, shift, std::forward<HandleEvent>(handle_event));
     }
 }
